Split StringFormatter into its own file
This fixes the checkstyle complain, and is also a good idea for reuse
with Bytes and ByteArray.
